split generic and specific tests
